requirements.txt
requests
pyyaml
watchdog
